add payment activity (#277)
