+++ demo/src/main/AndroidManifest.xml
@@ -29,7 +29,11 @@
             android:label="@string/custom_flow_title" />
 
         <activity
-            android:name="com.verygoodsecurity.vgscheckout.demo.payment.PaymentCheckoutActivity"
+            android:name="com.verygoodsecurity.vgscheckout.demo.orchestrtion.addcard.AddCardCheckoutActivity"
+            android:label="@string/add_card_flow_title" />
+
+        <activity
+            android:name="com.verygoodsecurity.vgscheckout.demo.orchestrtion.payment.PaymentCheckoutActivity"
             android:label="@string/payment_flow_title" />
 
         <activity
