+++ demo/src/main/java/com/verygoodsecurity/vgscheckout/demo/MainActivity.kt
@@ -4,8 +4,9 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.button.MaterialButton
-import com.verygoodsecurity.vgscheckout.demo.payment.PaymentCheckoutActivity
 import com.verygoodsecurity.vgscheckout.demo.custom.CustomCheckoutActivity
+import com.verygoodsecurity.vgscheckout.demo.orchestrtion.addcard.AddCardCheckoutActivity
+import com.verygoodsecurity.vgscheckout.demo.orchestrtion.payment.PaymentCheckoutActivity
 
 class MainActivity : AppCompatActivity(R.layout.activity_main) {
 
@@ -13,6 +14,9 @@ class MainActivity : AppCompatActivity(R.layout.activity_main) {
         super.onCreate(savedInstanceState)
         findViewById<MaterialButton>(R.id.mbStartCustom).setOnClickListener {
             startActivity(Intent(this, CustomCheckoutActivity::class.java))
+        }
+        findViewById<MaterialButton>(R.id.mbStartAddCard).setOnClickListener {
+            startActivity(Intent(this, AddCardCheckoutActivity::class.java))
         }
         findViewById<MaterialButton>(R.id.mbStartPayment).setOnClickListener {
             startActivity(Intent(this, PaymentCheckoutActivity::class.java))
