+++ demo/src/main/java/com/verygoodsecurity/vgscheckout/demo/orchestrtion/OrchestrationCheckoutActivity.kt
@@ -1,4 +1,4 @@
-package com.verygoodsecurity.vgscheckout.demo.payment
+package com.verygoodsecurity.vgscheckout.demo.orchestrtion
 
 import android.os.Bundle
 import android.os.Handler
@@ -6,23 +6,14 @@ import android.os.Looper
 import android.util.Log
 import android.view.View
 import android.widget.ProgressBar
-import androidx.preference.PreferenceManager
 import com.google.android.material.button.MaterialButton
 import com.google.gson.Gson
 import com.google.gson.annotations.SerializedName
+import com.verygoodsecurity.vgscheckout.BuildConfig.AUTHENTICATION_HOST
 import com.verygoodsecurity.vgscheckout.VGSCheckout
 import com.verygoodsecurity.vgscheckout.VGSCheckoutCallback
-import com.verygoodsecurity.vgscheckout.config.VGSCheckoutAddCardConfig
-import com.verygoodsecurity.vgscheckout.config.ui.VGSCheckoutAddCardFormConfig
-import com.verygoodsecurity.vgscheckout.config.ui.view.address.VGSCheckoutBillingAddressVisibility
-import com.verygoodsecurity.vgscheckout.config.ui.view.address.VGSCheckoutPaymentBillingAddressOptions
-import com.verygoodsecurity.vgscheckout.config.ui.view.address.address.VGSCheckoutPaymentAddressOptions
-import com.verygoodsecurity.vgscheckout.config.ui.view.address.address.VGSCheckoutPaymentOptionalAddressOptions
-import com.verygoodsecurity.vgscheckout.config.ui.view.address.city.VGSCheckoutPaymentCityOptions
-import com.verygoodsecurity.vgscheckout.config.ui.view.address.code.VGSCheckoutPaymentPostalCodeOptions
-import com.verygoodsecurity.vgscheckout.config.ui.view.address.country.VGSCheckoutPaymentCountryOptions
+import com.verygoodsecurity.vgscheckout.config.core.CheckoutConfig
 import com.verygoodsecurity.vgscheckout.demo.BaseActivity
-import com.verygoodsecurity.vgscheckout.demo.BuildConfig
 import com.verygoodsecurity.vgscheckout.demo.CheckoutType
 import com.verygoodsecurity.vgscheckout.demo.R
 import com.verygoodsecurity.vgscheckout.model.VGSCheckoutResult
@@ -32,7 +23,7 @@ import okhttp3.*
 import okhttp3.internal.EMPTY_REQUEST
 import java.io.IOException
 
-class PaymentCheckoutActivity : BaseActivity(R.layout.activity_payment_checkout),
+abstract class OrchestrationCheckoutActivity : BaseActivity(R.layout.activity_payment_checkout),
     VGSCheckoutCallback {
 
     override val type: CheckoutType = CheckoutType.PAYMENT
@@ -69,37 +60,16 @@ class PaymentCheckoutActivity : BaseActivity(R.layout.activity_payment_checkout)
                 Log.d(this::class.simpleName, "Token is null.")
                 return@get
             }
-            checkout.present(createConfig(it.value))
+            initializeConfiguration(it.value) { config ->
+                checkout.present(config)
+            }
         }
     }
 
-    private fun createConfig(token: String): VGSCheckoutAddCardConfig {
-        val preferences = PreferenceManager.getDefaultSharedPreferences(this)
-        val formConfig = VGSCheckoutAddCardFormConfig(
-            VGSCheckoutPaymentBillingAddressOptions(
-                VGSCheckoutPaymentCountryOptions(
-                    visibility = preferences.getFieldVisibility(R.string.setting_key_country_visible)
-                ),
-                VGSCheckoutPaymentCityOptions(preferences.getFieldVisibility(R.string.setting_key_city_visible)),
-                VGSCheckoutPaymentAddressOptions(preferences.getFieldVisibility(R.string.setting_key_address_visible)),
-                VGSCheckoutPaymentOptionalAddressOptions(preferences.getFieldVisibility(R.string.setting_key_optional_address_visible)),
-                VGSCheckoutPaymentPostalCodeOptions(preferences.getFieldVisibility(R.string.setting_key_postal_code_visible)),
-                if (preferences.getBoolean(R.string.setting_key_billing_address_visible)) {
-                    VGSCheckoutBillingAddressVisibility.VISIBLE
-                } else {
-                    VGSCheckoutBillingAddressVisibility.HIDDEN
-                }
-            ),
-            preferences.getValidationBehaviour(R.string.setting_key_validation_behaviour),
-            preferences.getBoolean(R.string.setting_key_save_card_option_enabled)
-        )
-        val vaultId = preferences.getString(getString(R.string.setting_key_vault_id), null)
-        return VGSCheckoutAddCardConfig(
-            accessToken = token,
-            tenantId = vaultId ?: BuildConfig.STORAGE_ID,
-            formConfig = formConfig
-        )
-    }
+    internal abstract fun initializeConfiguration(
+        token: String,
+        callback: (config: CheckoutConfig) -> Unit
+    )
 
     private fun setLoading(isLoading: Boolean) {
         mbPresent.isEnabled = !isLoading
@@ -115,7 +85,7 @@ class TokenManager {
 
     private val tokenRequest: Request by lazy {
         Request.Builder()
-            .url("https://multiplexing-demo.verygoodsecurity.io/get-auth-token")
+            .url(AUTHENTICATION_HOST)
             .post(EMPTY_REQUEST)
             .build()
     }
