+++ demo/src/main/res/values/strings.xml
@@ -1,10 +1,12 @@
 <resources>
     <string name="main_activity_title">VGS Checkout Demo</string>
     <string name="custom_flow_title">VGS Custom Config Demo</string>
-    <string name="payment_flow_title">VGS Payment Config Demo</string>
+    <string name="add_card_flow_title">VGS Add Card Demo</string>
+    <string name="payment_flow_title">VGS Payment Demo</string>
     <string name="settings_title">Checkout settings</string>
 
     <string name="present_custom_checkout_btn_title">Custom</string>
+    <string name="present_add_card_checkout_btn_title">Add Card</string>
     <string name="present_payment_checkout_btn_title">Payment</string>
 
     <string name="custom_checkout_present_btn_title">Present</string>
